// src/api/security/model/userInfoDTO.ts
/**
 * OpenAPI definition
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface UserInfoDTO { 
    id: string;
    sub: string;
    name: string;
    email: string;
    roles: Array<string>;
    email_verified: boolean;
    phone_number: string;
    phone_number_verified: boolean;
    locale: string;
    zoneinfo: string;
    updated_at: string;
}